requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.1
django-allauth==0.51.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.0
filelock==3.8.0
idna==3.3
mysqlclient==2.1.1
oauthlib==3.2.0
pipenv==2022.8.31
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
pyotp==2.6.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.4
virtualenv-clone==0.5.7
